[#66434] Update scimitar.
- Use custom_authenticator provied by scimitar to do authentication.
- Polish specs a bit.
- Rework OpenProject scimitar modifications structure.
